initial chunk sending is nearly working
had some showstoppers with me screwing up the rectangle positioning, still another bug where the chunk top-left starts at player position (should be centered on it)

also had a showstopper with kryonet's read buffer. apparently it does not auto resize its serialization buffers. 32768 (2^15) should be okay..the previous object size was 2^14-1000 or so

we may want to set up a queuing mechanism later, as 15,000 bytes is quite a bit (mtu is 1500)