hotbarinventory serialization over network and loading etc..
caught a few bugs along the way, big ones especially one that was hidden but could cause a player to try and spawn twice..
left to do is drag & drop of items and the network backing for that rearrangement. that's just for the hotbar for now.

the hotbar items server-side, stack size gets randomized as well, just to make it different.