improve the grass growth
also enable now sending of sparse block updates. instead of sending a single block,
for out-of-order (nonlinear regions) block updates, we can now send sparse block regions.
iow, when we need to send random blocks that need to get updated, but of course that
should be queued. it can probably help ensure good RLE, and is obviously more efficient
than either sending a single block or sending a range with excepted blocks.

of course, less efficient than a simple begin x, y, end x, y, as we need to send an index
with each block.

also touch up the world viewport code. though it is horribly broken but i realize now that i should not
improve it, until i get chunk loading, which will literally negate all of the  code here so far,
but be more efficient in a lot of areas...and easier to do checks like "look for chunks nearest me",
so we can trick the player into doing random things (like grass growth) in such areas.

it'd also have the concept of loadedness, which is necessary for > 1 player.

of course, it would also get rid of the horrible memory footprint on both sides (esp singleplayer)..

it would *also* allow us to do all-chunk sending, and incremental chunk updates (sparse blocks help this too)

in other words, i could stream chunks going to the right, as i move over them..instead of the wasteful
"send all the tiles ever because he moved a bit"