world about 65% converted to artemis, and some methods it calls into others
course systems need to be converted, and we need to use those.

i also need to reimplement component cloning in a different manner, seeing
as i cannot create them through manual instantiation, also i need a way
to query a list of entities with criterion and stuff..for isPlacementValid
collision detection. but perhaps this should actually reside in a system.