fix a lot more build errors, introduce inputsystem
not sure if it's a good way of handling stuff, but we need
a few diff types of input events..polling and event keys.

but polling needs to be done per tick, which can only be
done cleanly inside a system. ergo, this class...
some things may not yet necessarily fit anywhere cleanly, either.