profiler seems to work okay
though right now it is  only for client. i'd need to do quite a few changes
to make it apply to the server ...but i think i have to.

it'll need a few touchups in the gui i think. it also needs to learn
how to exclude *itself* from the profiler