fix power circuit tests, try to get some logic going for removing
a wire via the mouse..but the test is failing, so check to see if it
is beneath the mouse must be wrong.

also move the crosshair, itemplacementoverlay code into entityoverlaysystem
i'm not yet sure what this sytem will have in it, but for starters we
could just have these 2 things..if more becomes an issue we can break it out.

though, i broke the build so that i remember to fix this method, because
it needs to get cleaned up and broken into other methods, like.

right now it's very hard to read.