floor rendering positions of sprites, fixes issue where
dropped items would, when the player moves, wiggle about.
but placed items would not. that was because the placed
items were being floored in logic, because they
want to be aligned. but dropped items just dropped
whereever they wanted. this fixes that, thankfully.

add debug code for getting entity at position,
so i can list stats and stuff in the debug text overlay.