improve debug text rendering, show a right text panel for info on entity
that is under mouse.

also fix various querying around here for finding list of pl ayers, remove
the manually maintained list. ECS handles this.

cleanup debugtext system and a lot of dead code in movement system

the start of trying to make how to determine what the server/client
is running under and what the underlying connection is as well.
e.g. if it's only a server connection. also, ability to determine
what this precise server instance is. who owns it.