start of workarounds for efficiently storing blocks in memory
because java.

essentially, since we have 4 bytes (the primitive) of data per block to store,
we will instead store each one in one byte[], so every 4 bytes there will be a
block. we'll need to overhaul our block retrieval/modification stuff,
and introduce a bunch of helper methods to do so, because obviously it's
going to make things trickier otherwise.