create circuitid, wireid, we're going to serialize these with
every entity's devicecomponent. this way the server and client each
know what we're talking about.

i'm thinking that the client will rarely have a complete view of all
the circuits and wires in a circuit. so, the server will actually
have to send them updates when stats change (like total supply/demand)