add some circuit, wire identifiers to devices
realized i need a list in the devicecomp, of wireid's. not my preference,
given that gets serialized. but i don't want to preoptimize.
i used to have a reference, owningCircuit, but seeing as we're using
wire and circuit id's, and we need to serialize that over anyways, and
references here can't be, and they are never used for fast retrieval anyways
(nor would they need to ever be, likely), then this is better.