initial multithreaded approach :)
right now it only uses 2 threads, gonna split it up nicer to make it use
n threads.

now i have an intermediate float data array, which it partitions evenly
for each thread(2 for now), then at the end join()'s them up and transfers
to png and writes it. was having issues earlier when i was writing directly
to the png..which was strange, because i wasn't crossing partitions in the
image data array...and the underlying code i didn't see anything that would
disagree with being reentrant. but, ended up getting noise (undesireable noise)
in the center border between the partitions. so that was interesting.

but yeah, right now a simple heightmap with mountains, valleys etc, takes
46 seconds with 1 thread on my cpu. with 2 it's about 24 seconds. and no
foreseeable issues.

of course, we create noise modules for each thread. but that's actually very
lightweight from what i see (cpu and memory wise). it's just when you query
them, that is the expensive part.

this'll get much more expensive when i combine fractals, for caves and stuff
and then finally, ore generation (long ways away) and other stuff.