create another test function to generate 500 random worlds, output to png
will help testing a lot, as i can narrow down which ones are problematic.

also, the seed used to generate it is baked onto the image, so i can use
that to track it down

rename block fields, put the word byte in them so it's more clear, there are
only so far 4 bytes per block..so far.