fix progression of ores..the value is super important,
because if we want a common to be more common than a rare,
we need the high/low source to be right, which implies
everything has to be done in the right order and with the right
progressing values (shoving diamond right in the middle of
the selection means everything following it is heavily skewed
against)

gold isn't so super common. of course, now coal is actually
super common. not sure how i feel about that.