add iron to the noise selections...
add color as well, for it.

also tweak coal and some other ores, course now that these
are tweaked, gold and everything else isn't going to appear.
but we want coal not to be *super* common so you find it
every 5 steps. and we want copper to be maybe a bit more rare
than that, and iron might be a bit more.