first add of attempts at light level code (sunlight)
i know why the flood fill is failing, but i haven't yet uncovered the
proper fix for this. unfortunately it isn't a simple "start at y=0 and
continue downward until you hit ground and then start attenuating".

because that isn't what terraria does even. terraria seems to actually
bleed light into caves. so e.g. a cave that is cut into the mountain,
can would be completely dark through the other mechanism. but through
terraria's mechanism makes it so that a large portion of the cave,
depending on how close to the entrance which is exposed to sunlight is,
it will be most lit. seems most people online aren't aware of that when
they go to implement this..