lighting works!
although it's only at initial world gen time. anytime after that,
i'll have to think of how best to do it. i'm not sure if i can just trigger
flood fill on each one that gets modified, or if that'd break..everything
i'm thinking i may have to update an entire section of blocks when 1
or so is modified, not sure.