cleanup peak finding function
would like to get rid of that hashmap that the original code thought
they needed for some strange reason.

but now we can, depending on delta, identify what kinda valleys and mountains
we are aware of. that's not possible otherwise, since we were using implicit
methods of terrain generation.

so it's a second pass which must be done.

but we've safely identified most of the ones i'd be interested in.

i may want separate passes, one to identify the smaller ones, another
to identify larger ones, and then we can do what we want based on that.