make inventories smarter, they should now be able to figure out where
the next free spot is and use it, they should also be able to
merge/combine items by identifying if they are of the same type
(slightly complex since they're all using components and i have to
identify each one and ensure it's the same).

also abstract away the details of the item component record keeping.
that should be handled by helpers in the inventory themselves.

yet untested, but the end result so far is the server spawn code
is much much cleaner. though all of that is debug code anyways. but this
will be used for item pickups, rewards and things like that