start of creating fuel source functionality, specifically slots
for the generator to have to be able to burn fuel.

i suppose the server will keep the an Inventory in each
generatorcomponent, client can get those serialized over
when they request to open that window.

i think i want a subclass though, because otherwise i have to treat
the first index or two as a slot in the array, which is awful.

like how mc has it, which is a pile of fuel slots and
1 fuel that's being burnt