fix duplication for different types of inventory spawning
guess i'll just go with 1 generic packet, even if some of the stuff
that is in there will not be set according to it. i could improve
this down the road if i wanted to, more easily as well.

if it mattered..which it probably doesn't, not exactly high throughput