fix some handling here
now it at least breaks in a visible way, instead of a silent/do nothing way.
had to add some convenience functions around inventory, so we don't
have to worry about indices as much. for a lot of things that's more a pain,
especially when you can just filter on the element and now we return a
reference.