more checks to handle some odd cases for liquid
allow placing blocks on top of liquid, thus clearing/getting rid of the liquid

still seems to be at least one issue, where water poured in from the top
into a pool will kind of lag in settling in, almost as if it's viscous.

i suspect this is because i only allow 1 actual processing to happen each step,
so it doesn't move too invisibly quickly..not sure yet

also make entity factory, which is just a class that will
have prefabs in it, instead of where they were in the world.