asynchronously launch world gen. this isn't the right solution
but it is the path to one. it's actually kind of cool because you spawn
in the world instantly and see the world change around you. unfortunately,
that isn't going to work later on, because there are certain processes and
such that form and deform around you, that would be disruptive/probably kill
you. plus npcs and things like that....

i could have decided to just generate the portion you're in and do the rest
in the background, but again..not worth it because several processes happen
at different times and that's just not worth the headache of bugs from them.

still, a very cool bug right now. but really what it should do is just be async,
in the sense that it is gen'ing while we are staring at a loading screen
(aka the render thread continues so we can show progress).