initial attempt at adding canCombineWith to each component
via the component interface. we'll see if this works at all...things will
break for sure, right away.

this all stems from the issue of the inventory trying to combine items,
and not doing so, due to me not updating the functino for each new component
i add. very troublesome. there's also another place which is in the copying,
i'm hoping to improve both of those conciseness and automaticness-wise

also override copying functions for every comp, some that should've..didn't
now they all have to, at least.